DataCollection/requirements.txt
spotipy==2.23.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
jupyter==1.0.0 